Started with CRA and implemented the Porjectbar, set up the basic styling
The ProjectBar is changed slightly so it doesn't throw an error (caused by me giving the React.Fragment a className-Attribute). The Project-Design is similar to the other FCC-Front-end-Libaries-Project. I simply centered the userinterface in the middle of the screen, the ProjectBar can be found on the left side. For the first time I decided to try out glassmorphism. Due to its accessibility-disadvantages, I'am not shure to use it in an larger project. So this mini project is a good test area for that.